requirements.txt
flask
cs50
python-dotenv
flask-session
openai
#google-generativeai
